+++ package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "bootstrap-vue": "^2.0.0-rc.11",
+    "firebase": "^5.5.2",
     "material-design-icons-iconfont": "^3.0.3",
     "vue": "^2.5.2",
     "vue-router": "^3.0.1",
+    "vuefire": "^1.4.5",
     "vuetify": "^1.2.3"
   },
   "devDependencies": {
